requirements.txt
tensorflow>=2.19.0
numpy>=1.26.0
opencv-python>=4.10.0
matplotlib>=3.9.0
tqdm>=4.66.0
absl-py>=2.1.0
